Refactor some classes into Utilties package
There are certain classes that were previously not classified in any
specfic package. These classes included all the utilities related to
Duke, such as the Storage, UI etc..

Refactoring these classses improves the overall quality of code.
