+++ src/main/java/duke/Command/ToDoCommand.java
@@ -2,12 +2,12 @@ package duke.Command;
 
 import duke.Task.ToDo;
 
-import duke.Storage;
-import duke.TaskList;
-import duke.UI;
+import duke.Utilities.Storage;
+import duke.Utilities.TaskList;
+import duke.Utilities.UI;
 
 /**
- * The ToDo command which is executed by duke.Duke.
+ * The ToDo command which is executed by duke.Utilities.Duke.
  */
 public class ToDoCommand extends Command {
