+++ src/main/java/duke/Exception/DukeException.java
@@ -1,12 +1,12 @@
 package duke.Exception;
 
 /**
- * The abstraction behind each exception thrown by duke.Duke.
+ * The abstraction behind each exception thrown by duke.Utilities.Duke.
  */
 public abstract class DukeException extends Exception {
 
     /**
-     * The constructor for a duke.Duke exception.
+     * The constructor for a duke.Utilities.Duke exception.
      * @param message The error message.
      */
     public DukeException(String message) {
