+++ src/main/java/duke/Utilities/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.Utilities;
 
 import java.io.IOException;
 import java.util.Collections;
