+++ src/main/java/duke/Utilities/UI.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.Utilities;
 
 import java.util.Scanner;
 
@@ -31,10 +31,10 @@ public class UI {
     }
 
     /**
-     * Prints to the UI an error message, if duke.Duke is being run for the first time and the required data files to
+     * Prints to the UI an error message, if duke.Utilities.Duke is being run for the first time and the required data files to
      * read from do not exist.
      */
     public void showLoadingError() {
-        System.out.println("Running duke.Duke for the first time - Creating required data files");
+        System.out.println("Running Duke for the first time - Creating required data files");
     }
 }
